fixed hypervisor table sort redirection to engines
Fixed faulty sorting on hypervisor selection page.
Moved shared sorting and filtering functionality to mixin.
