+++ fusor-ember-cli/app/controllers/engine/discovered-host.js
@@ -1,8 +1,12 @@
 import Ember from 'ember';
+
 import NeedsDeploymentMixin from "../../mixins/needs-deployment-mixin";
 import PaginationControllerMixin from "../../mixins/pagination-controller-mixin";
+import FilterSortHostsMixin from "../../mixins/filter-sort-hosts-mixin";
 
-export default Ember.Controller.extend(NeedsDeploymentMixin, PaginationControllerMixin, {
+export default Ember.Controller.extend(NeedsDeploymentMixin, PaginationControllerMixin, FilterSortHostsMixin, {
+
+  sortRoute: "engine.discovered-host",
 
   rhevController: Ember.inject.controller('rhev'),
 
@@ -40,32 +44,6 @@ export default Ember.Controller.extend(NeedsDeploymentMixin, PaginationControlle
       return !isHypervisor && !isDeploying;
     });
   }),
-
-  filteredHosts: Ember.computed('availableHosts.[]', 'searchString', 'isStarted', function(){
-    var searchString = this.get('searchString');
-    var rx = new RegExp(searchString, 'gi');
-    var availableHosts = this.get('availableHosts');
-
-    if (this.get('isStarted')) {
-      return Ember.A([this.get('model')]);
-    } else if (availableHosts.get('length') > 0) {
-      return availableHosts.filter(function(record) {
-        return (record.get('name').match(rx) || record.get('memory_human_size').match(rx) ||
-                record.get('disks_human_size').match(rx) || record.get('subnet_to_s').match(rx) ||
-                record.get('mac').match(rx)
-               );
-      });
-    } else {
-      return availableHosts;
-    }
-  }),
-
-  sortCriteria: Ember.computed('sort_by', 'dir', function () {
-    let sort_by = this.get('sort_by') || 'name';
-    let dir = this.get('dir') || 'asc';
-    return [sort_by+':'+dir];
-  }),
-  sortedHosts: Ember.computed.sort('filteredHosts', 'sortCriteria'),
 
   numSelected: Ember.computed('model.id', function() {
     return (this.get('model.id')) ? 1 : 0;
